requirements.txt
aiogram>=3.4
python-dotenv>=1.0
